requirements.txt
# Core ML and API frameworks
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# Data processing and visualization
numpy>=1.24.0
pillow>=9.5.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0

# Geospatial data handling
rasterio>=1.3.0
shapely>=2.0.0
pyproj>=3.5.0

# Utilities
pydantic>=2.0.0
pathlib
typing-extensions
requests>=2.31.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
